Moving the base-path extraction to utils.py.
